Fix Safari Reader mode dropping initial paragraphs.
When there are multiple sub-sections (level3 divs), paragraphs that
appear between section header and those subsections are dropped by
Safari Reader mode. Wrapping the paragraphs in a div that marks them as
a section of content causes Safari to properly include them in Reader
mode.

For more information, see this Webkit bug report:
https://bugs.webkit.org/show_bug.cgi?id=226853